cidaasnative/src/main/java/de/cidaas/sdk/android/cidaasnative/data/entity/register/registrationsetup/RegistrationSetupAttributesEntity.java
package de.cidaas.sdk.android.cidaasnative.data.entity.register.registrationsetup;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

import java.io.Serializable;

@JsonIgnoreProperties(ignoreUnknown = true)
public class RegistrationSetupAttributesEntity implements Serializable {
    String key;
    String value;

    public String getKey() {
        return key;
    }

    public void setKey(String key) {
        this.key = key;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }
}
